Split tests into separate files for upgrades/upgradevalidation package;
